requirements.txt
asgiref==3.3.4
Django==3.2.3
django-crispy-forms==1.11.2
django-model-utils==4.3.1
django-multiselectfield==0.1.12
django-notifications==0.1.dev0
django-notifications-hq==1.8.3
jsonfield==3.1.0
notifications==0.3.2
Pillow==10.1.0
pytz==2021.1
sqlparse==0.4.1
swapper==1.3.0
whitenoise==6.6.0
